Remove the Add Shape Docker
This docker is extremely dangerous these days because it can
be used to create ODF based shapes that cannot be saved or
copied. Note that when loading Krita 3 files with vector layers
that contain vector shapes or text shapes or other shapes for
which we don't have a clone() methods defined, this assert

SAFE ASSERT (krita): "0 && "not implemented!"" in file /krita-4.0.0.51/libs/flake/KoShape.cpp, line 315

Will still fire: we need to make sure this cannot happen before
we release.

CCBUG:389271