Improve Python-> Scripter's exception reporting
Currently scripter posts a critical QMessage box containing the error
text relating to an exception, but it does not provide the line number
or text where the exception occurred. This requires someone scripting
to close the message box before correcting the issue.
For syntax errors and indent errors it will provide the line number but
not (in the case of syntax errors) any visual indication of the
location of the syntax error.

This diff takes exception handling into runaction.py and provides
exception reporting in the output widget. This removes the dialog box,
so script developers don't need to constantly close a dialog box to
edit and rerun their script. The reporting is minimal, effectively
limited to one traceback.

For syntax errors it shows the location of the error.

Differential Revision: https://phabricator.kde.org/D9531