Add auto-indent functionality to Python Scripter editor
The return key has just been pressed (and processed by the editor)
now insert leading spaces to reflect an appropriate indent level
against the previous line.
This will depend on the end of the previous line. If it ends:

*  with a colon (:) then indent to a new indent level
*  with a comma (,) then this is an implied continuation line, probably
in the middle of a function's parameter list look for last open bracket
on previous line (, [ or { if found indent to that level + one
character, otherwise use previous line whitespace, this is probably a
list or parameter list so line up with other elements
*  with a backslash (\) then this is a continuation line, probably on
the RHS of an assignment
*  similar rules as for comma, but if there is an = character use that
plus one indent level if that is greater if it is an open bracket of
some sort treat similarly to comma

*  anything else - a new line at the same indent level. This will
preserve the indent level of whitespace lines. User can shift-tab to
dedent as necessary

Differential Revision: https://phabricator.kde.org/D9445