Show a warning when trying to paint on a clone layer
This shows the same warning as when trying to paint with the
freehand tool on a vector layer. Rectangle etc. tools also
should show the warning, but don't yet.

Note that this isn't a complete solution, but generalizing
the message would need changes to the strings, and we're in
string freeze now.

Another option would be to enable/disable tools dependent
on the type of layer selected, but that is a big refactoring,
since tools only react to the type of shape selected at the
moment.

CCBUG:389570