Remove VectorShape
It was not used anyway :) After the refactoring of KisMimeDatabase,
the database started to return 'application/x-what' for the files without
any extension, therefore vector shape was not used for loading them.
Instead, KoUnavailShape was used to fetch the data.

Now embedded SVG files are loaded directly by
KoShapeRegistry::createShapeFromOdf() so a special shape is not needed
anymore.

NOTE: we dropped the support of WMF, EMF and other file types previously
      handled by VectorShape!