Fix updates of shape layers when the image in transformed with size change
The problem was:

The shape layer crops the prerendered projection by the image size. And
during the transformation steps the image resize happened **after** the
all the shape updates has been completed. It meant that these "new" areas
of the image were never updated

The patch also implements a correct updates compression in the shape
layers and its handling by
KisDelayedUpdateNodeInterface::forceUpdateTimedNode()

BUG:381944