Check all changed node properties when determining undo/redo
If there's even one property in the list of properties that has
changed and that isn't one of the properties that should not end
up in the undo list, we should make an undo command.

And if there isn't any, then then the document should not be set
to modified.

BUG:389876