Scratchpad: Implement "Fill Layer"
So, I implemented this because the button for it had shown up but didn't do anything.

The text for this said "TODO" and given the way it was named(fillLayer)
and the fact that the fill with background color was implemented just fine,
I figured the intention was to have the current image be copied into the scratchpad.

So that is what this does.

CCMAIL:Kimageshop@kde.org