Brush Settings: Adjust live-preview to work better with the sketchbrush.
It now draws something like a little sinusoid, and it takes into account that
the sketch brush shouldn't have it's size limited.

This also sets the curve and particle brushes to use the little wave because they
give better previews for that. I also suspect the bristle brush engine might benefit from
this, but maybe keeping it on the simple line might be better.