Rewritten painting of handles in KoCreatePathTool
Now the handles are painted in a uniform way by a couple of
KisHandlePainterHelper + KoShapeHandlesCollection. KoPathPoint
generates an abstract list of "handles", which are sorted in the
tool and painted in one go. It allows the tool to decide, which
area on screen the handle occupies and issue updateCanvas() calls
accordingly.

CCBUG:381913