CPMT: Add support for exporting panels and text.
It's fairly rudimentary, but it beats transcribing by hand by a mile.

This supports the strong and emphasis tags, but not yet the strokethrough one.
Nor is there support for background color or type. Not even sure how I should
handle the last.

To have vector layers that use panels or text be parsed, they need to contain
"text" or "panels" in the name. I will make this configurable in the future.