Add pager docker
This docker allows paging through documents contained in the same
folder.

When a page is turned, the active document is closed and the
(alphabetically) next/previous .kra file is opened in its place. If
multiple files from the same directory are open, all of them will be
paged together by stepping the corresponding number of files ahead or
back.