Windows: Reset screen rotation after probing QPA OpenGL
If the Windows QPA OpenGL blacklist has disabled rotation for the
device, it does not reset the screen rotation preference and as a
result the screen orientation will remain locked in landscape even if
ANGLE is enabled or the QPainter canvas is used.

This change works around it by resetting the screen rotation after
the QPA OpenGL probing. The display might still be forced into landscape
for a split second during startup, but it is better than being locked
entirely.

CCBUG: 390651