Fix Crashing when Changing Smoothing Option in a Stroke (Bug 386726)
The stablizer draws lines even after user lifts his pen, but when
the smoothing option is set to other options, the related information
are immediately deleted, so when the stabilizer tries to continue to
draw the lines, a crash is caused.

Also, the stablizer's timer continues to run after the smoothing
option is changed, and draws lines when timeout; this causes regression
in lines.

Differential Revision: https://phabricator.kde.org/D10157