Scripting: add View::selectedNodes
This returns a list of nodes selected in the given View. Note
that it is possible for different views to have a different
set of selected nodes, so being selected is not a property of
the node.

This also adds a utility class for libkis, so we can share
some common code that doesn't need to be public.

BUG:390818