Live Brush Preview: Limit more paintop settings to reasonable sizes.
When having massive texture sizes, or massive brushes on the spraybrush,
the gui will freeze updue to live preview having trouble with all this.

This may seem like a case where the user shoots themselves in the foot
by picking too big sizes, but when using it, it isn't obvious that the
paintop preview is causing the GUI freeze, and it is more common to cause
these freezes by fatfingering controls than it is to cause freezes while
painting with a slow brush. Therefore let's limit these values to reasonable
sizes.

BUG:390897