Update brush-preset thumbnail: Eraser_Screentone.
The preset was looking alike Eraser_Circle when zoom out.
I decorated the preset with a blue sand eraser (often the blue-side
of pink-blue eraser). This type can remove layer of material 
on the surface of the paper. Sometime it works for ink. I saw it
used for degrading screentones dots on the plastic before applying them.
I also added a 'screentone' symbol on the corner.