Update brush-preset thumbnail: Deform family (Grow,Move,Shrink)
The three presets represented only a wireframe grid with the deformation.
While this system was pretty good when thumbnail was tiny, it wasn't 
aesthetic at all in the brush selection.  I couldn't reuse the style in my own 
brushkit because I needed three variations. 

Here is what I've done:
[Grow] arrows are in diagonal (as in liquify icon), color is green meaning + 
and silhouette is a big ball
[Move] arrows are perpendicular, and silhouette is like a directional gamepad
color is blue (to split from Grow and Shrink)
[Shrink] color is orange/redish, as a star collapsing, arrows are diagonal too
for same reason than Grow, a circle with dashed line show the previous state 
and suggest the movement of Shrinking.

All three digital-stylus tools where slightly colored to also reflect the color of 
the actions.