Update brush-preset thumbnail: Watercolor and Waterpaint family
A nice trio of presets; not a real simulation of watercolors, but they 
help to match this texture of watercolor edges, fringe and drop of
waters. A powerful set.

I decided to use green for the stroke family (in my brushkit, I already use this
color for Watercolors).  The brush body is cyan and have written on it  "Aqua"
to simulate a well known brand of watercolor brushes.  The presets to do
texturing has dark hair on the tip. (Petit Gris)  with a puffy shape. The two other
one to paint shape and details have synthetic yellow tips.