Update brush-preset : Airbrush Soft
Replace the default of Krita with the one on my brushkit.
Mine had less artifact and better rendering and no issue
with Instant-Preview. The performance of mine are probably
less good (even if I use precision 1) because I'm using a lower
spacing by default. But the quality of an airbrush is in how 
subtle it can be. Krita already struggle to get a good and smooth
fading of airbrush tip, so it's better to not screw this preset.