Update brush-preset thumbnail: revert tone down background color
After a short-experiment and getting used to the b4b4b4 grey color
I found it was looking a bit sad in flat variant. Sure, it offers
a bit more room for value, and I don't have to cap my value to white
but it also looks boring.
I'm reverting to use cecece grey background ( the default for 4.x )