Update brush-preset thumbnail: Eraser (full set)
I moved Eraser_Screentone to Screentone_Eraser;
I think this tool is more specific to a part for screentones,
grouped together, the Screentones preset will make sens for
someone who test the brushes.

I adopted the Circle and Soft of 4.x as my default. Eraser Hard
Wasn't really helpfull in anyway. Aliased, no dynamic. I replaced
it with my Eraser_Small to assist artist in erasing fully lines, and 
details. Also, the brush tip is squary, so it impossible to be confused
when one use a workflow of brush plus eraser and using swith to
previous preset.