Update brush-preset thumbnail: Ink (partial set)
Two presets in 4.x were similar to two of my brushkit:
- Ink Tought ( Fill brush on my side )
- Ink Precision  (annotation brush on my side).
I merged ( I kept the settings of 4.x and decorated the presets
with icon of my brushkit ) because the settings in 4.x were 
superior ; more precise, better curve and optimised.

Ink Leaky has now a feather artwork and Ink Fineliner has a 
tubular pen with a 0.5 and red coating.