Refactor brush-preset:  Stamp Mountains (kpp+gih)
An animated brush part of my Chaos&Evolution DvD, 2009.
I redid the animated brush ( 4 random cell ), with mountains
I painted for the Spring project. I upgraded the mask size to 
256x256 for performances ; and the kpp get a new icon and 
a settings with more spacing and large size (360px) to not lag
on small machine. The brush have been tested for Instant Preview
too to be WYSIWYG.