Add color themes: Krita darker
Not a new one, from 2012, Krita 2.5 [1], the theme was a bit broken;
The selection color was a bright grey, and not the default Krita-blue,
Also the GUI had too much contrast. I fixed it from the viewpoint of
my system. I'm not sure how it behaves in Windows or Mac. 

[1] http://www.davidrevoy.com/article124/themes-for-krita-and-kde