Update all thumbnails and all presets and merge my Brushkit8.2
This big commit replace the previous 4.x presets with a new version. 90% of the 
previous presets still exists under a new name or a new appearance. My 
Brushkit 8.2 was merged into the list and I made a triage of the duplicate. I changed
mostly the setting to optimize the presets: for getting good size "out-of-the-box"
on canvas at 33% or 50% zoom of the viewport, I also tweaked the settings to
get better performances. I also changed the brush doing glitches or bad rendering
when the opacity slider was changed, or the brush forcing too much on the blending
mode. I also changed the pressure curves of the brushes to get similar response to 
pressure ( many presets were customised to the tablet built-in pressure of their owner).
I tested here on three tablets with various pressure curves ( intuos3, intuos4, and cintiq13hd).
I used at least a preset for each brush engine of Krita ; and renamed the linked new brushes 
bitmap-tip necessary for merging my brushkit. The kit is also all sorted. I'll keep testing during the
next week and I'll probably refine it until the release deadline.

Note: I made a quick bundle of the previous version :
https://www.peppercarrot.com/extras/resources/2018-03-02_4xalpha-replaced.bundle