Add three new presets: Screentones, Impressionism, Shapes Square
Screentones regular is a brush designed to add screentones in a
classic way; regular pattern. The pattern density can be adjusted
with the brightness of the texture, or replaced with another
pattern in the library, or provided by the user.

Impressionism is a brush designed to add a lot of stroke with HSV
variation and obtain a simili Vincent van Gogh effect as in the
famous "The Starry Night" painting (1889).

Shapes Square is a non-realistic and pure digital brushes, not
trying to mimic any traditional tool. It was designed to showcase
how Krita brush engine can be flexible and self-suffisent to do
assumed "computer-art" style.

(cherry picked from commit c1be1c881ce15c315ce96b0b012e50f28fef88cc)