Refactor brush presets for performances: heavy textured brushes
A couple of presets were really slow. Mainly because activating all the
bells and whistles of recents features like brush-mask. (eg. using
brush-mask as a way to texture the stroke, while using texture is more
efficient in terms of performances ). Also, Some stroke produced
good looking strokes, but after painting tests, were totally unpredictable,
slow, felt a bit duplicate of other texture effect and were hard to control.

Sometime, sensors were added as "rotation" on "drawing direction" on both
brush and brush-mask (forgoting extra "rotation on pressure"). A sign a lot
of this brushes were designed a bit by pressing all the buttons. It's fine
for a creative process, but needs cleaning. Especially while the benefit
of the resulting effect is not justifying such a cost on perfomances.

I tried to re-adapt the brushes and extract the target effect while redoing
them in different way, but more predictable for the end user using it, and
also with less impact on the performances.