Add tags for brush list
Warning: two tags refuse to display , named "Ink" and "FX"
I think the strings are sort of blacklisted (from 3.x ?)

The list of tag is simple and list preset for suggesting categories
to the final user:

- Digital:
List all presets with a rendering only digital tools can does.

- Paint:
List a suggestion of presets to make textured/traditional paint.

- Ink:
Tools for the black&white artist (comic, manga, illustration).

- Pixel-Art:
Three tools for the pixel artist

- Sketch:
A list of preset to put down quickly an idea on the canvas. Contain
also all charcoal/pastel/pencil and other dry techniques.

- Textures
All the stamps and patterns in the library.

- My Favourites
A list of ten brush to plug in the right-click pop-up palette by default
and a generic name to encourage the end user to use it for his/her own
brushes.