Set Default.gpl for color selector + adapt layout to 10 column.
The default brushkit was created for 10 columns but adapted to
Krita 20 column at a time the palette docker could only display
fixed size little square, and 20 column per row. Now the palette
is doing a liquid rescale to the number of column. An opportunity
to switch it back to 10 columns!

Also, the color selector now will use default instead of falling
back to the first *.gpl palette found in the folder
(the cookie-palette.gpl). It's easier to pick a base color and all
the color are named in this palette.