Fix copy-pasting of shapes after changing image resolution
Steps to reproduce:

1) Create a rectangular shape
2) Image->Scale Image: do not rescale the image, just change its
   resolution, say, from 300 ppi to 600 ppi.
3) Select the rectangle, press Ctrl+C, Ctrl+V

The shape should be pasted in the original size.

The problem it that when changing image resolution we scale the parent
shape layer, but the shape itself is not changed. Therefore, when copying
the shape into the clipboard, we should also save the shape's absolute
transformation, so that it could be pasted correclty.

(cherry picked from commit 82280e2a3c33fe1d60cdbe75d83bdd965a427df2)