Fix resetting LoD options when changing the brush settings
After we started to reset all the old brush settings before writing,
saving of LoD settings broke. The problem is that LoD widget is outside
of the brush settings widgets hierarchy, therefore, it is not written
an every change. Instead, the value of the option is handled by the
resource manager.

The proposed solution is a bit "hackish", but I have no idea
how to fix it without a full rewrite of the paintop editor :(

BUG:391050