Fix loading the action plugins in new windows
We were caching the loaded plugins, so a new window would not load
the per-window action plugins, which made for empty menus.

However, the Python plugin was a KisActionPlugin but should only
be loaded once, on application start, when the first window is
being created. It should not be reloaded when a new window is
created.

(cherry picked from commit ab523d80132e086e71cf40332c540bc71c55cee6)