Load the tags in a separate step after loading all resources
This also refactors KoTagStore to remove two methods the use
of which has disappeared in the mists of time, and that
prevented some tags to be loaded for default resources.

Now all Deevad's tags are loaded; all system tags used for
custom resources in 3.3 are kept; all custom tags created
in 3.3 are kept.

The regression from the previous commit, where the krita 3
bundle was loaded even if disabled is also fixed...

Items 1, 2 and 4 of 391576 are resolved now, at least as
far as I can see. I haven't dared look at item 3 yet...

CCBUG:391576