Paintoppreset: improvement for i)_Wet_Bristles
Changes sent by Radian over IRC, it improves the smoothness of the
brush but impact the performance. But all in all the rendering is
better this way and does a better work at being a "Wet_Bristles"
presets.

This commit update tags md5, delete previous SVG brushes and
replace it with the GBR provided by radian.

main changes:
- from 80 to 90px radius
- from 0.5 to 0.3 spacing
- brushmask is GBR now and 300px radius, instead of SVG 1000px radius.