Try harder to setup the environment for finding translations
We can only get the application root if we have an application, so
create a dummy temporary qapplication to get the root; then set
all the paths and environment, then create the real application,
then set the application domain. Now ki18n should be able to find
the translations under all circumstances.

CCBUG:392044