Do not try to infinitely iterate through all the pixels of the device if its extent is empty
For historical reasons we use weird format for "empty extent" mark,
the rectangle is set to QRect(-int32max, -int32max, 0, 0). Using
such mark is a bit of dangerous, because mere operating with it
causes integer overflow and, therefore, infinite loop. I'm not sure
if it is safe to change this mark into standard Qt's one...

BUG:384587