Make the text editor window application modal.
You should not even be able to select tools or do anything with any
object on the canvas while the text editor window is open. However,
that only works if the editor window has a parent, and it cannot have a
parent, because it has a menubar, so if it would have a parent, it would
crash on macOS. Setting it to ApplicationModal would make it block all
windows...

Still, that's better than allowing users to select other tools while
the window is active.

BUG:392248
(cherry picked from commit ee806c8304125f7a88a04ecd6058cf41932090be)