On switching a workspace unlock all dockers
The workspace selector widget tried to restore the lock status
after changing a workspace for all visible dockers, the workspace
menu didn't do that. In any case, it is far too easy to lock
invisible dockers, and that confuses far too many people and takes
far too much of our time to explain.

Tt is better to define that changing a workspace will reset the
lockedness of all dockers. The lock property of a docker is not
part of the workspace.

BUG:392235
CCMAIL:kimageshop@kde.org

Maybe we should remove the lock and collapse feature for dockers
altogether.