Do not save and load channel lock flags to .kra
Until the channel docker was made to work on the projection in

    commit 365d0eb775cbff3bf6afda3f9b703741dbd68710
    Author: Boudewijn Rempt <boud@valdyas.org>
    Date:   Mon Sep 30 09:02:41 2013 +0200

    Start making the channel docker work on the projection, not the layer

We could lock channels and painting on them would be disabled. Except
for some remnants in libqml, that is no longer possible. However, we
would still load and save the channel lock flags. Maybe in the future
we should add the channel lock flags to the layer properties dialog
instead and restore this code.

I'm still not sure how the file in the bug report was created, unless
it originated in Krita Gemini.

BUG:392504
(cherry picked from commit 7a27f04cbe093ff1880a261f38504c2dfa0800d1)