Remove the lock, collapse and hide option for dock titlebars
On asking around, it seems that even Krita power users were not
using these options. New users to Krita seemed unable to grasp this
functionality altogether. And we were always having trouble managing
the locked and collapsed state of dockers.

Note that not having a custom titlebar widget also makes it impossible
to keep having an option to remove the titlebars altogether, which was
another way to force the dockers to stay in one place.

A side-effect is that on macOS, floating dockers are resizable now.

BUG:385238
CCBUG:392538
CCMAIL:kimageshop@kde.org