Add script starter script
Adds a new script, the krita script starter. The script, when run, will
prompt the user for meta data related to the script, then create the
file structure to get started on a the script, primarily based on the
name given to the script. This makes it easy to create new scripts - eg
for new users, or for creating scripts for testing purposes.

Differential Revision: https://phabricator.kde.org/D11275