Disable handling Ctrl+Z as "cancel" action in KisToolMove
Basically, we shouldn't cancel the stroke on undo in "continued"
tool actions. Instead the user should use "Esc" key for that.

Otherwise the previous command (the one that came before the
current action) will be undone, which is not what the user wants.

Ideally we should finally implement an undo system for the move tool,
but it is still a todo.

CCBUG:392014