Disable right-clicking on popup palette
With this patch it is not possible to select a palette item
using the right-click of the stylus. All the distinct actions
should be confirmed with the left button. Right-click on any part
of the canvas just closes the entire palette.

BUG:391696,378484
(cherry picked from commit 20c9efffb70cf2757595fee8d7c56a5fe59d5664)