Fix copy-paste frames operation when the timeline layout changes
Now "application/x-krita-frame" mime type also saves the UUID of the
source node, so pasting the frames after the row layout changed is still
valid.

The patch also implements a new version of
KisAnimationUtils::createMoveKeyframesCommand which is safe to use
when the frame moves have cyclic dependencies (one frame is going to
overwrite another frame during the move). Theoretically, we don't need
to use sortPointsForSafeMove() anymore. But it should be tested.