Rename KisAirbrushOption to KisAirbrushOptionWidget
This makes clear that this class is a widget and cannot be created
during painting. Now we'll have to split the widget part and the
settings part up. We probably should rename all classes that inherit
KisPaintOpOption like this, to make the difference clear between
the KisPaintOpOption hierarchy. the KisBaseOption hierarchy and the
KisCurveOption hierarchy.

CCMAIL:allen.marshall.22@gmail.com
CCBUG:392190