Disable QtQuick on macOS
The touch docker isn't that awesome on systems without a touch
screen and running the qml plugin saves cached compiled files
in the app bundle and that breaks signing. So disable this
until I've found out how to make Qt Quick on macOS save its
generated files in the user's home folder.

CCBUG:394188