Implement a basic Frame Cache Swapping
Right now the frames are saved with LZF compression over
frame differences. It makes compression ratio about 10x
times higher, but the speed (of LZF decompression) is still
rather slow.

A FullHD frame (1920x1080) is decompressed in about 16ms, which
is fine, but if the user creates a 6k image, then playback will
simply be impossible, because LZF compression will take about
60ms of time.