Fix loading of the Python library
Summary:
Instead of assuming the unversioned .so symlink will exist also at
runtime (which is not the case for binary distros), use the QLibrary
features: the SONAME of the Python library is "1.0", so use the
unsuffixed filename of Python library with it.

Additionally, this extends the loading also to Unices platforms
different than macOS (since Boud excluded it in a previous commit),
and destroy the QLibrary in case of loading failure.

BUG: 393307
Fixed-In: 4.0.4

Reviewers: rempt

Reviewed By: rempt

Subscribers: rempt

Differential Revision: https://phabricator.kde.org/D13144

(cherry picked from commit 1553c12637e15d77168e6ddad287ccbf0c53075b)