Adjust similarity Tolerance.
We no longer adjust for a percentage, this can cause the test to fail
in cases where the difference is huge in one or two pixels. Not acceptable
for a painting mask. (White dot in the middle of mask would have passed test)

Diference between pixels can be at most 2 values apart, and
no error is tolerated for the test to pass