Copy bundles to the local bundle folder from args
If Krita is started with a bundle file as an argument, do not open
it, but copy it to the bundle folder. You can also drop a bundle
on the mainwindow but it won't be loaded in that case.

BUG:394485
(cherry picked from commit 8fff40a222b006f211e907800b15fe035005a8f6)