Optimize Fill Layers to not regenerate stuff when not needed
With this patch one can use Fill Layer as an overlay for
colored draft lines. Painting works almost at the same speed
as painting on a normal layer.

Such usecase was suggested by David Revoy at Krita Sprint 2018

To Deevad: it would be nice to have a short video about that :)

CC:kimageshop@kde.org