Try to make Krita run as an appimage again
This reverts part of

commit 6509644e3cb89610fce1f2925b0126c9bc1439c0
Author: Jeroen Hoolmans <jhoolmans@gmail.com>
Date:   Wed May 16 15:29:35 2018 +0200

    Added support to compile with Python2.7 on Linux

The code in this block was never executed because of a wrong
case-sensitive check for the mount point of an appimage. Now
it runs, it crashes all the time on *buntu, and sometimes
on opensuse.

CCMAIL:jhoolmans@gmail.com
(cherry picked from commit 410255d029cf4ceaadc4d95ef0c9edefb32f471c)