Cleanup the python path setting
This puts back the old way of setting the paths. I'm not sure
whether this will give a build error with python 2; if so, it
needs to be put between ifdefs, but I'm not sure which ifdefs
are available. In any case, this block of code is only relevant
for appimages, and since we're not providing python 2.7 appimages,
the code would't be executed by 2.7 using krita builds.

CCMAIL:jhoolmans@gmail.com
BUG:395606
(cherry picked from commit e6568f4a9a8ce26248d4e2f310911ffbc80ff3ad)