Fix assert when trying to merge down a group with a transform mask
We should be very careful about passing node pointers in queued signals.
Sometimes it may happen that the signal is delivered *after* the node has
been removed from the layers graph. If it is possible and we should avoid
that, e.g. activating this node, we should check node's graph listener's
pointer.

BUG:395003