Remove the recorder
We've been disabling this in releases since forwever because itr was
never even nearly finished: for instance, the wdgpaintactioneditor form
was never used anywhere, most tools didn't implement the special
recording code, all of our actions were never recorded.

A design where extra code has to be added in multiple places for every
single thing that we want to record isn't going to work.

If we want to do a recorder, we should start thinking about how we can
do this in a generic way, by having the UI generate xml or json that gets
recorded and executed: I guess we'd need to add some kind of interpreter
at the heart of Krita. But this is a dead end...

CCMAIL:kimageshop@kde.org