pykrita: expose the real i18n() function
Summary:
Using QCoreApplication.translate() as i18n() function in pykrita does
not work, since the translations are not in Qt .qm files, but in
Gettext .po files.

Hence, create an helper krita_i18n() function in libkis, bind it with
SIP, and set it as builtin i18n() function for pykrita (and thus for
all the Python plugins).  This fixes most of the lack of translations
in Python plugins.

Test Plan:
Builds fine, and Python plugins are translated properly
(mostly, since there are few more issues, but unrelated to this).

Reviewers: #krita, rempt

Reviewed By: #krita, rempt

Subscribers: rempt

Tags: #krita

Differential Revision: https://phabricator.kde.org/D13689