Fix PyKrita library search path for MacOS
If Krita detects that its embedded Python comes from a system
installation, it must preserve the original search path (Py_GetPath)
and append the PyKrita module paths.

This prevents PyKrita from not being able to find the 'sip' and
'encodings' modules, which in turn cause the Python engine to
fail to initialize and to outright crash, respectively.

Differential Revision: https://phabricator.kde.org/D13682